refactor(core-featureflag): change to kmp
