+++ app-common/build.gradle.kts
@@ -17,7 +17,7 @@ dependencies {
     implementation(projects.legacy.core)
     implementation(projects.legacy.account)
 
-    implementation(projects.core.featureflags)
+    implementation(projects.core.featureflag)
     implementation(projects.core.ui.legacy.theme2.common)
 
     implementation(projects.feature.account.setup)
