+++ app-k9mail/build.gradle.kts
@@ -142,7 +142,7 @@ dependencies {
     implementation(projects.legacy.core)
     implementation(projects.legacy.ui.legacy)
 
-    implementation(projects.core.featureflags)
+    implementation(projects.core.featureflag)
 
     implementation(projects.feature.account.settings.impl)
