+++ app-k9mail/src/debug/kotlin/app/k9mail/featureflag/K9FeatureFlagFactory.kt
@@ -1,8 +1,8 @@
 package app.k9mail.featureflag
 
-import app.k9mail.core.featureflag.FeatureFlag
-import app.k9mail.core.featureflag.FeatureFlagFactory
-import app.k9mail.core.featureflag.toFeatureFlagKey
+import net.thunderbird.core.featureflag.FeatureFlag
+import net.thunderbird.core.featureflag.FeatureFlagFactory
+import net.thunderbird.core.featureflag.toFeatureFlagKey
 
 class K9FeatureFlagFactory : FeatureFlagFactory {
     override fun createFeatureCatalog(): List<FeatureFlag> {
