+++ app-k9mail/src/main/kotlin/app/k9mail/K9KoinModule.kt
@@ -2,7 +2,6 @@ package app.k9mail
 
 import app.k9mail.auth.K9OAuthConfigurationFactory
 import app.k9mail.core.common.oauth.OAuthConfigurationFactory
-import app.k9mail.core.featureflag.FeatureFlagFactory
 import app.k9mail.dev.developmentModuleAdditions
 import app.k9mail.feature.featureModule
 import app.k9mail.feature.widget.shortcut.LauncherShortcutActivity
@@ -15,6 +14,7 @@ import com.fsck.k9.activity.MessageCompose
 import com.fsck.k9.provider.UnreadWidgetProvider
 import com.fsck.k9.widget.list.MessageListWidgetProvider
 import net.thunderbird.app.common.appCommonModule
+import net.thunderbird.core.featureflag.FeatureFlagFactory
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
