+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/ThunderbirdKoinModule.kt
@@ -1,7 +1,6 @@
 package net.thunderbird.android
 
 import app.k9mail.core.common.oauth.OAuthConfigurationFactory
-import app.k9mail.core.featureflag.FeatureFlagFactory
 import app.k9mail.feature.widget.shortcut.LauncherShortcutActivity
 import com.fsck.k9.AppConfig
 import com.fsck.k9.activity.MessageCompose
@@ -14,6 +13,7 @@ import net.thunderbird.android.widget.provider.MessageListWidgetProvider
 import net.thunderbird.android.widget.provider.UnreadWidgetProvider
 import net.thunderbird.android.widget.widgetModule
 import net.thunderbird.app.common.appCommonModule
+import net.thunderbird.core.featureflag.FeatureFlagFactory
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
