+++ core/featureflag/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id(ThunderbirdPlugins.Library.kmp)
+}
+
+android {
+    namespace = "net.thunderbird.core.featureflag"
+}
