+++ core/featureflag/src/commonMain/kotlin/net/thunderbird/core/featureflag/FeatureFlag.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.featureflag
+package net.thunderbird.core.featureflag
 
 data class FeatureFlag(
     val key: FeatureFlagKey,
