+++ core/featureflag/src/commonMain/kotlin/net/thunderbird/core/featureflag/FeatureFlagFactory.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.featureflag
+package net.thunderbird.core.featureflag
 
 fun interface FeatureFlagFactory {
     fun createFeatureCatalog(): List<FeatureFlag>
