--- core/featureflags/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    id(ThunderbirdPlugins.Library.jvm)
-    alias(libs.plugins.android.lint)
-}
-
-dependencies {
-    testImplementation(projects.core.testing)
-}
