+++ legacy/core/src/main/java/com/fsck/k9/K9.kt
@@ -2,8 +2,6 @@ package com.fsck.k9
 
 import android.content.Context
 import android.content.SharedPreferences
-import app.k9mail.core.featureflag.FeatureFlagProvider
-import app.k9mail.core.featureflag.toFeatureFlagKey
 import app.k9mail.feature.telemetry.api.TelemetryManager
 import app.k9mail.legacy.account.AccountDefaultsProvider
 import app.k9mail.legacy.account.SortType
@@ -15,6 +13,8 @@ import com.fsck.k9.mailstore.LocalStore
 import com.fsck.k9.preferences.RealGeneralSettingsManager
 import com.fsck.k9.preferences.StorageEditor
 import kotlinx.datetime.Clock
+import net.thunderbird.core.featureflag.FeatureFlagProvider
+import net.thunderbird.core.featureflag.toFeatureFlagKey
 import net.thunderbird.core.preferences.Storage
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
