+++ legacy/core/src/test/java/com/fsck/k9/TestApp.kt
@@ -2,9 +2,6 @@ package com.fsck.k9
 
 import android.app.Application
 import androidx.work.WorkManager
-import app.k9mail.core.featureflag.FeatureFlag
-import app.k9mail.core.featureflag.FeatureFlagProvider
-import app.k9mail.core.featureflag.InMemoryFeatureFlagProvider
 import app.k9mail.feature.telemetry.telemetryModule
 import app.k9mail.legacy.account.AccountDefaultsProvider
 import app.k9mail.legacy.di.DI
@@ -17,6 +14,9 @@ import com.fsck.k9.notification.NotificationStrategy
 import com.fsck.k9.preferences.StoragePersister
 import com.fsck.k9.storage.storageModule
 import net.thunderbird.core.android.preferences.InMemoryStoragePersister
+import net.thunderbird.core.featureflag.FeatureFlag
+import net.thunderbird.core.featureflag.FeatureFlagProvider
+import net.thunderbird.core.featureflag.InMemoryFeatureFlagProvider
 import net.thunderbird.legacy.core.FakeAccountDefaultsProvider
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
