+++ legacy/storage/build.gradle.kts
@@ -16,7 +16,7 @@ dependencies {
     testImplementation(projects.feature.telemetry.noop)
     testImplementation(libs.robolectric)
     testImplementation(libs.commons.io)
-    testImplementation(projects.core.featureflags)
+    testImplementation(projects.core.featureflag)
 }
 
 android {
