+++ legacy/storage/src/test/java/com/fsck/k9/storage/TestApp.kt
@@ -1,9 +1,6 @@
 package com.fsck.k9.storage
 
 import android.app.Application
-import app.k9mail.core.featureflag.FeatureFlag
-import app.k9mail.core.featureflag.FeatureFlagProvider
-import app.k9mail.core.featureflag.InMemoryFeatureFlagProvider
 import app.k9mail.feature.telemetry.telemetryModule
 import app.k9mail.legacy.account.AccountDefaultsProvider
 import app.k9mail.legacy.di.DI
@@ -16,6 +13,9 @@ import com.fsck.k9.crypto.EncryptionExtractor
 import com.fsck.k9.legacyCoreModules
 import com.fsck.k9.preferences.K9StoragePersister
 import com.fsck.k9.preferences.StoragePersister
+import net.thunderbird.core.featureflag.FeatureFlag
+import net.thunderbird.core.featureflag.FeatureFlagProvider
+import net.thunderbird.core.featureflag.InMemoryFeatureFlagProvider
 import org.koin.dsl.module
 import org.mockito.kotlin.mock
