+++ legacy/ui/legacy/src/test/java/com/fsck/k9/TestApp.kt
@@ -1,15 +1,15 @@
 package com.fsck.k9
 
 import android.app.Application
-import app.k9mail.core.featureflag.FeatureFlag
-import app.k9mail.core.featureflag.FeatureFlagProvider
-import app.k9mail.core.featureflag.InMemoryFeatureFlagProvider
 import app.k9mail.feature.telemetry.telemetryModule
 import app.k9mail.legacy.account.AccountDefaultsProvider
 import app.k9mail.legacy.di.DI
 import com.fsck.k9.contacts.ContactPictureLoader
 import com.fsck.k9.preferences.StoragePersister
 import net.thunderbird.core.android.preferences.InMemoryStoragePersister
+import net.thunderbird.core.featureflag.FeatureFlag
+import net.thunderbird.core.featureflag.FeatureFlagProvider
+import net.thunderbird.core.featureflag.InMemoryFeatureFlagProvider
 import org.koin.dsl.module
 import org.mockito.Mockito.mock
